backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
pydantic[email]==2.10.1
pydantic-settings==2.6.1
bcrypt==4.0.1
PyJWT==2.10.1
python-multipart==0.0.12
slowapi>=0.1.9

openai>=1.0.0
httpx>=0.27.0
pymupdf4llm>=0.0.17
fpdf2>=2.8.0
